# backend/docs/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for CI/CD application developed using Go (Gin, Gorm).",
        "title": "CI/CD Application - Server API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/projects": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get all projects",
                "operationId": "all-projects",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Filter",
                        "name": "project",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.ProjectCore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of projects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProjectShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Error in request",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Create new project",
                "operationId": "create-project",
                "parameters": [
                    {
                        "description": "New project entry",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProjectCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Newly created project",
                        "schema": {
                            "$ref": "#/definitions/model.Project"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get the single project",
                "operationId": "single-project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requested project",
                        "schema": {
                            "$ref": "#/definitions/model.Project"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No record found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Update project",
                "operationId": "update-project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated project entry",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProjectCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated project",
                        "schema": {
                            "$ref": "#/definitions/model.Project"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Delete project",
                "operationId": "delete-project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete message",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/secrets": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Get all secrets",
                "operationId": "all-secrets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Filter",
                        "name": "secret",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.SecretCore"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of secrets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SecretShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Error in request",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "404": {
                        "description": "No records found",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Create new secret",
                "operationId": "create-secret",
                "parameters": [
                    {
                        "description": "New secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Newly created secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
        "/secrets/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Update secret",
                "operationId": "update-secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Secret ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated secret entry",
                        "name": "secret",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SecretCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated secret",
                        "schema": {
                            "$ref": "#/definitions/model.Secret"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Delete secret",
                "operationId": "delete-secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Secret ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete message",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "400": {
                        "description": "Error in params",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    },
                    "501": {
                        "description": "Endpoint not implemented",
                        "schema": {
                            "$ref": "#/definitions/util.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Build": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStep"
                    }
                },
                "tags": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
        "model.BuildLog": {
            "type": "object",
            "properties": {
                "build_step_id": {
                    "type": "string"
                },
                "command": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                }
            }
        },
...
        "model.BuildStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "BuildRunning",
                "BuildSuccessful",
                "BuildFailed",
                "BuildCanceled"
            ]
        },
        "model.BuildStep": {
            "type": "object",
            "properties": {
                "build_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Pipeline": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "builds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Build"
                    }
                },
                "config": {
                    "$ref": "#/definitions/model.PipelineConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "secrets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Secret"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Variable"
                    }
                }
            }
        },
        "model.PipelineConfig": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "commands": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "name": {
                                "type": "string"
                            }
                        }
                    }
                },
                "system": {
                    "type": "string"
                }
            }
        },
...
        "model.Project": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pipelines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Pipeline"
                    }
                },
                "repo": {
                    "type": "string"
                },
                "secrets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Secret"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Variable"
                    }
                }
            }
        },
        "model.ProjectCore": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                }
            }
        },
        "model.ProjectCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                }
            }
        },
        "model.ProjectShort": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Secret": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SecretCore": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                }
            }
        },
        "model.SecretCreate": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.SecretShort": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Variable": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        },
        "util.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}